Move arc_dijkstras::Graph to a separate file
